+++ internal/plancal/root.go
@@ -55,6 +55,7 @@ var rootCmd = &cobra.Command{
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute(version string) {
+	config.Set(config.Version, version)
 	rootCmd.Version = version
 	err := rootCmd.Execute()
 	if err != nil {
@@ -66,5 +67,4 @@ func init() {
 	rootCmd.PersistentFlags().StringVar(&logLevel, "loglevel", "info", "")
 
 	logging.InitLogging(logLevel, true)
-	config.InitConfig()
 }